requirements.txt
django>=5.0,<6.0
django-allauth>=0.58.0
requests>=2.31.0
PyJWT>=2.8.0
cryptography>=42.0.0
django-environ>=0.11.2
dj-database-url>=2.1.0
gunicorn>=22.0.0
psycopg2-binary>=2.9.9
whitenoise>=6.6.0
django-storages>=1.14.2
boto3>=1.34.78
